ASC-109 Test nova create
This commit adds a test to create a nova (compute) resource and ensures
that it can be connected to.

This test relies on the previous deployment of openstack api access as
well as network, image, flavor, security groups, and ssh key resources.
Therefore, this commit also includes a converge task pulling in
[ansible-role-pytest-rpc](https://github.com/rcbops/ansible-role-pytest-rpc).
and running the role.